requirements.txt
Flask
Flask-SQLAlchemy
Flask-Minify
Flask-Bootstrap
Flask-Cors
Flask-Login
Flask-Migrate
Flask-WTF
python-decouple
python-dotenv
psycopg2-binary
email-validator
pytest
flake8